feat: Rotas autorizadas com jwt e oauth do github.

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { FastifyInstance } from 'fastify';
 import { prisma } from '../lib/prisma';
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify();
+  });
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -17,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       }));
   });
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -27,6 +34,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     });
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) return res.status(401).send();
 
     return memory;
   });
@@ -45,14 +54,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '1ea30006-5d3a-4004-bfb7-c748dfb25fef',
+        userId: req.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put('/memories/:id', async (req) => {
+  app.put('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -67,7 +76,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(req.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== req.user.sub) return res.status(401).send();
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -79,12 +94,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete('/memories/:id', async (req) => {
+  app.delete('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(req.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== req.user.sub) return res.status(401).send();
 
     await prisma.memory.delete({
       where: { id },
